fix the upload

+++ src/api/property/property.ts
@@ -2,11 +2,13 @@
 
 
 import express, { Request, Response } from "express";
-import { PropertyImageInput, PropertyInput, propertySchema, PropertyTagInput, PropertyUpdateInput, propertyUpdateSchema } from "../../lib/property";
+import { ProjectUnitType, PropertyImageInput, PropertyInput, propertySchema, PropertyTagInput, PropertyUpdateInput, propertyUpdateSchema } from "../../lib/property";
 import prisma from "../../../prisma/prisma";
 import { authMiddleware } from "../../middleware/auth";
 import { subscriptionMiddleware } from "../../middleware/subscription";
 import { SubscriptionFeatures } from "../../lib/subscription";
+import slugify from "slugify";
+import { UserMiddleware } from "../../middleware/user-middleware";
 
 const router = express.Router();
 
@@ -22,7 +24,7 @@ router.post("/",
             }
 
             const { property_images, property_tags, project_units, ...propertInput } = result.data;
-
+            (propertInput as PropertyInput).slug = slugify(propertInput.title, { lower: true, strict: true });
             const createdProperty = await prisma.$transaction(async (tx) => {
 
                 return await tx.properties.create({
@@ -47,7 +49,7 @@ router.post("/",
                                 }
                             }))
                         }
-                    },
+                    } as any,
                     include: {
                         property_images: true,
                         property_property_tags: {
@@ -63,6 +65,7 @@ router.post("/",
 
             response.status(200).json({ success: true, createdProperty });
         } catch (error) {
+            console.log (error)
             response.status(500).json({ success: false, error });
         }
     });
@@ -102,17 +105,19 @@ router.put("/:id",
             }
 
             const { property_images, property_tags, project_units, ...propertyInput } = result.data;
-
+            if (propertyInput.title) {
+                (propertyInput as PropertyInput).slug = slugify(propertyInput.title, { lower: true, strict: true });
+            }
             const updatedProperty = await prisma.$transaction(async (tx) => {
 
-                const newImageIds = property_images.filter((img: PropertyImageInput) => img.id).map((img: PropertyImageInput) => img.id);
-                const imagesToDelete = property.property_images.filter((img) => !newImageIds.includes(img.id))
-                const updatedUnits = project_units.filter((unit) => unit.id);
-                const updatedUnitesIds = updatedUnits.map((unit) => unit.id)
-                const deletedUnites = property.project_units.filter((unit) => !updatedUnitesIds.includes(unit.id));
+                const imagesToKeep: number[] | undefined = property_images?.filter((img: PropertyImageInput) => img.id).map((img: PropertyImageInput) => img.id);
+                const imagesToDelete = imagesToKeep !== undefined ? property.property_images.filter((img) => !imagesToKeep.includes(img.id)) : undefined
+                const updatedUnits: ProjectUnitType[] | undefined = project_units?.filter((unit) => unit.id);
+                const updatedUnitesIds: number[] | undefined = updatedUnits?.map((unit) => unit.id)
+                const deletedUnites: ProjectUnitType[] | undefined = updatedUnitesIds !== undefined ? property.project_units.filter((unit) => !updatedUnitesIds.includes(unit.id)) : undefined;
 
                 // remove images 
-                if (imagesToDelete.length > 0)
+                if (imagesToDelete?.length > 0)
                     await tx.property_images.deleteMany({
                         where: {
                             id: {
@@ -121,20 +126,22 @@ router.put("/:id",
                         }
                     });
 
+                let property_property_tags: any[] = [];
+                if (property_tags !== undefined) {
 
-                // ---- Handle Tags (M2M) ----
-                // Clear old tags first (simpler and consistent)
-                await tx.property_property_tags.deleteMany({
-                    where: { property_id: id },
-                });
+                    // ---- Handle Tags (M2M) ----
+                    // Clear old tags first (simpler and consistent)
+                    await tx.property_property_tags.deleteMany({
+                        where: { property_id: id },
+                    });
 
-                // Recreate links (connect existing or create new)
-                const property_property_tags = property_tags.map(tag => ({
-                    property_tags: tag.id
-                        ? { connect: { id: tag.id } }
-                        : { create: { name: tag.name, user_id: (request as any).user.id } },
-                }));
-
+                    // Recreate links (connect existing or create new)
+                    property_property_tags = property_tags?.map(tag => ({
+                        property_tags: tag.id
+                            ? { connect: { id: tag.id } }
+                            : { create: { name: tag.name, user_id: (request as any).user.id } },
+                    })) || [];
+                }
                 return await tx.properties.update({
                     where: {
                         id
@@ -142,16 +149,16 @@ router.put("/:id",
                     data: {
                         ...propertyInput,
                         property_images: {
-                            create: property_images.filter((img) => !img.id).map((img) => ({ image_url: img.image_url }))
+                            create: property_images?.filter((img) => !img.id).map((img) => ({ image_url: img.image_url })) || []
                         },
                         property_property_tags: { create: property_property_tags },
                         project_units: {
-                            update: updatedUnits.map((unit) => ({
+                            update: updatedUnits?.map((unit) => ({
                                 where: { id: unit.id },
                                 data: unit
                             })),
-                            create: project_units.filter((unit) => !unit.id) as any,
-                            deleteMany: deletedUnites.map((unit) => ({ id: unit.id }))
+                            create: project_units?.filter((unit) => !unit.id) as any || [],
+                            deleteMany: deletedUnites?.map((unit) => ({ id: unit.id }))
                         }
                     },
                     include: {
@@ -173,8 +180,72 @@ router.put("/:id",
     });
 
 
+router.get("/area", async (request: Request, response: Response) => {
+    try {
 
-router.get("/:id", async (request: Request, response: Response) => {
+
+        const result = await prisma.properties.aggregate({
+            _min: { area_sq_meters: true },
+            _max: { area_sq_meters: true },
+        });
+
+        response.status(200).json({
+            success: true,
+            data: {
+                min_area: Number ( result._min.area_sq_meters) ,
+                max_area: Number ( result._max.area_sq_meters),
+            }
+        }) ; 
+        
+    } catch (error) {
+        response.status(500).json({ success: false, error })
+    }
+})
+
+
+router.get("/slug/:slug" , UserMiddleware as any , async (request: Request, response: Response) => {
+    try {
+        const slug = request.params.slug;
+        
+        
+        if (!slug) {
+            response.status(401).json({ success: false, error: "Slug is required" }); return;
+        }
+
+        const property = await prisma.properties.findUnique({
+            where: {
+                slug
+            },
+            include: {
+                property_images: true,
+                property_property_tags: {
+                    include: { property_tags: true },
+                },
+                
+                favorites: {
+                    where: {
+                        user_id: (request as any).user?.id
+                    },
+                },
+                users : {
+                    include : { social_media : true }
+                } , 
+                project_units: true
+            }
+        });
+        if (!property) {
+            response.status(404).json({ success: false, error: "Property not found" }); return;
+        }
+        response.status(200).json({ success: true, data: property });
+
+    } catch (error) {
+        console.log (error)
+        response.status(500).json({ success: false, error });
+    }
+});
+
+
+router.get("/:id" ,  UserMiddleware as any , async (request: Request, response: Response) => {
     try {
         const id = Number(request.params.id);
 
@@ -191,24 +262,25 @@ router.get("/:id", async (request: Request, response: Response) => {
                 },
                 favorites: {
                     where: {
-                        user_id: (request as any).user.id
+                        user_id: (request as any).user?.id
                     },
-                } , 
-                project_units : true 
+                },
+                project_units: true
             }
         });
         if (!property) {
             response.status(404).json({ success: false, error: "Property not found" }); return;
         }
-        response.status(201).json({ success: true, data: property });
+        response.status(200).json({ success: true, data: property });
 
     } catch (error) {
+        console.log(error);
         response.status(500).json({ success: false, error });
     }
 });
 
 
-router.get("/", async (request: Request, response: Response) => {
+router.get("/" ,  UserMiddleware as any , async (request: Request, response: Response) => {
     try {
 
         let {
@@ -228,6 +300,8 @@ router.get("/", async (request: Request, response: Response) => {
             favorite,
             latitude,
             longitude,
+            status,
+            user_id,
             sort_by = "created_at"
         } = request.query;
 
@@ -307,6 +381,18 @@ router.get("/", async (request: Request, response: Response) => {
             }
         }
 
+        if (status) {
+            filters = {
+                ...filters,
+                status
+            }
+        }
+
+        if (user_id) {
+            filters = {
+                user_id: JSON.parse(user_id as string)
+            }
+        }
         if (favorite == "true") {
             filters = {
                 ...filters,
@@ -359,10 +445,18 @@ router.get("/", async (request: Request, response: Response) => {
                     take: parseInt(limit as string),
                     include: {
                         property_images: true,
+                        users: {
+                            select: {
+                                id: true,
+                                full_name: true,
+                                picture_url: true
+                            }
+                        },
                         favorites: {
                             where: {
-                                user_id: (request as any).user.id
+                                user_id: (request as any).user?.id
                             },
+
                         }
                     },
                     orderBy: {
@@ -410,6 +504,8 @@ router.get("/", async (request: Request, response: Response) => {
 
             // ad type
             if (ad_type?.length) {
+
+
                 conditions.push(`p.add_type = ANY(${params.push(ad_type as string) && `$${params.length}`}::properties_add_type_enum[]) `)
             }
 
@@ -428,8 +524,12 @@ router.get("/", async (request: Request, response: Response) => {
             if (ownership_book !== undefined) conditions.push(`p.ownership_book = ${params.push(ownership_book) && `$${params.length}`}::boolean`)
 
             // favorites only
-            if (favorite === "true") {
-                conditions.push(`EXISTS (SELECT 1 FROM favorites f WHERE f.property_id = p.id AND f.user_id = ${params.push((request as any).user.id) && `$${params.length}`})`)
+            if (favorite === "true" && (request as any).user?.id) {
+                conditions.push(`EXISTS (SELECT 1 FROM favorites f WHERE f.property_id = p.id AND f.user_id = ${params.push((request as any).user?.id) && `$${params.length}`})`)
+            }
+
+            if (status) {
+                conditions.push(`p.status = '${status}'`)
             }
 
             // combine all
@@ -442,7 +542,7 @@ router.get("/", async (request: Request, response: Response) => {
             const lat = parseFloat(latitude as string);
             const lng = parseFloat(longitude as string);
 
-
+            console.log(whereClause)
             const [properties, countResult] = await prisma.$transaction([
                 prisma.$queryRawUnsafe(`
                     SELECT 
@@ -453,7 +553,7 @@ router.get("/", async (request: Request, response: Response) => {
                         sin(radians(${lat})) * sin(radians(p.latitude))
                         )) AS distance,
                         json_agg(DISTINCT pi.*) AS property_images,
-                        json_agg(DISTINCT f.*) FILTER (WHERE f.user_id = ${(request as any).user.id}) AS favorites
+                        json_agg(DISTINCT f.*) ${(request as any).user?.id ? `FILTER (WHERE f.user_id = ${(request as any).user?.id}) AS favorites` : ""}
                     FROM properties p
                     LEFT JOIN property_images pi ON pi.property_id = p.id
                     LEFT JOIN favorites f ON f.property_id = p.id
@@ -475,8 +575,7 @@ router.get("/", async (request: Request, response: Response) => {
 
             response.status(200).json({ success: true, count, data: properties }); return;
         }
-    } catch (error) {
-        console.log(error)
+    } catch (error) { 
         response.status(500).json({ success: false, error });
     }
 });
@@ -503,7 +602,7 @@ router.delete("/:id",
             } catch (error) {
                 response.status(404).json({ success: false, error: "Property not found" }); return;
             }
-            response.status(201).json({ success: true, data: id }); return;
+            response.status(200).json({ success: true, data: id }); return;
 
         } catch (error) {
             response.status(500).json({ success: false, error });
@@ -558,7 +657,6 @@ router.put("/favorite/:id",
             response.status(500).json({ success: false, error });
         }
     });
-
